+++ src/components/ui.js
@@ -88,7 +88,9 @@ export default class InspectorUI extends Component {
 	render() {
 		if ( this.state.isCollapsed ) {
 			document.body.classList.remove( 'ck-inspector-body-expanded' );
+			document.body.classList.add( 'ck-inspector-body-collapsed' );
 		} else {
+			document.body.classList.remove( 'ck-inspector-body-collapsed' );
 			document.body.classList.add( 'ck-inspector-body-expanded' );
 		}
 
@@ -99,6 +101,7 @@ export default class InspectorUI extends Component {
 			enableResizing={{ top: !this.state.isCollapsed }}
 			disableDragging={true}
 			minHeight={INSPECTOR_MIN_HEIGHT}
+			maxHeight="100%"
 			style={INSPECTOR_STYLES}
 			className={[
 				'ck-inspector',
@@ -129,6 +132,11 @@ export default class InspectorUI extends Component {
 				<CommandsPane label="Commands" editor={currentEditorInstance} />
 			</Tabs>
 		</Rnd>;
+	}
+	
+	componentWillUnmount() {
+			document.body.classList.remove( 'ck-inspector-body-expanded' );
+			document.body.classList.remove( 'ck-inspector-body-collapsed' );
 	}
 
 	static getDerivedStateFromProps( props, state ) {