+++ tests/inspector/components/propertylist.js
@@ -43,4 +43,23 @@ describe( '<PropertyList />', () => {
 		expect( dt1.find( 'label' ) ).to.have.attr( 'for' ).equal( dd1.find( 'input' ).prop( 'id' ) );
 		expect( dt2.find( 'label' ) ).to.have.attr( 'for' ).equal( dd2.find( 'input' ).prop( 'id' ) );
 	} );
+
+	it( 'truncates property values to 2000 characters', () => {
+		const items = [
+			[ 'foo', new Array( 1999 ).fill( 0 ).join( '' ) ],
+			[ 'bar', new Array( 2000 ).fill( 0 ).join( '' ) ],
+			[ 'baz', new Array( 2100 ).fill( 0 ).join( '' ) ]
+		];
+
+		wrapper = mount( <PropertyList items={items} /> );
+
+		const dd1 = wrapper.children().childAt( 1 );
+		const dd2 = wrapper.children().childAt( 3 );
+		const dd3 = wrapper.children().childAt( 5 );
+
+		expect( dd1.find( 'input' ).props().value ).to.have.length( 1999 );
+		expect( dd2.find( 'input' ).props().value ).to.have.length( 2000 );
+		expect( dd3.find( 'input' ).props().value ).to.have.lengthOf.below( 2100 );
+		expect( dd3.find( 'input' ).props().value ).to.match( /characters left]$/ );
+	} );
 } );